fix permission issue

+++ volatile_test/test.c
@@ -2,7 +2,7 @@
 
 int test_volatile(void)
 {
-	volatile uint32_t *reg_a = (uint32_t)0x40023800;
+	volatile uint32_t *reg_a = 0x40023800;
 
 	uint32_t val1;
 	uint32_t val2;
